+++ packages/web-vue/components/_components/dropdown/utils.ts
@@ -1,14 +1,4 @@
-import { VNode, isVNode } from 'vue';
 import { isArray, isFunction, isObject } from '../../_utils/is';
-import {
-  getBooleanProp,
-  getChildrenFunc,
-  getVNodeChildrenString,
-  isArrayChildren,
-  isNamedComponent,
-  isSlotsChildren,
-  resolveProps,
-} from '../../_utils/vue-utils';
 import {
   Option,
   OptionData,
@@ -23,7 +13,6 @@ const isGroupOption = (option: Option): option is GroupOption => {
 };
 
 export const getOptionNodes = ({
-  children,
   options,
   extraOptions,
   inputValue,
@@ -31,7 +20,6 @@ export const getOptionNodes = ({
   optionInfoMap,
   enabledOptionSet,
 }: {
-  children?: VNode[];
   options?: Option[];
   extraOptions?: Array<string | number>;
   inputValue?: string;
@@ -40,51 +28,33 @@ export const getOptionNodes = ({
   enabledOptionSet: Set<string | number>;
 }) => {
   const nodes: OptionNode[] = [];
+  optionInfoMap.clear();
+  enabledOptionSet.clear();
 
   const getAndSaveOptionInfo = (
-    option: string | number | OptionData | VNode,
-    origin: 'children' | 'options' | 'extraOptions'
+    option: string | number | OptionData,
+    origin: 'options' | 'extraOptions'
   ): OptionInfo | undefined => {
     const index = optionInfoMap.size;
 
-    let optionInfo: OptionInfo;
-    if (isVNode(option)) {
-      const props = resolveProps(option);
-      const label = props.label ?? getVNodeChildrenString(option);
-      const value = props.value ?? label;
-      const render = getChildrenFunc(option);
-      const key = `option-${typeof value}-${value}`;
-
-      optionInfo = {
-        ...props,
-        index,
-        key,
-        value,
-        label,
-        render,
-        disabled: getBooleanProp(option.props?.disabled),
-        origin: 'children',
-      };
-    } else {
-      optionInfo = isObject(option)
-        ? {
-            ...option,
-            index,
-            key: `option-${typeof option.value}-${option.value}`,
-            value: option.value,
-            label: option.label ?? String(option.value),
-            disabled: Boolean(option.disabled),
-            origin,
-          }
-        : {
-            index,
-            key: `option-${typeof option}-${option}`,
-            value: option,
-            label: String(option),
-            disabled: false,
-            origin,
-          };
-    }
+    const optionInfo: OptionInfo = isObject(option)
+      ? {
+          ...option,
+          index,
+          key: `option-${typeof option.value}-${option.value}`,
+          value: option.value,
+          label: option.label ?? String(option.value),
+          disabled: Boolean(option.disabled),
+          origin,
+        }
+      : {
+          index,
+          key: `option-${typeof option}-${option}`,
+          value: option,
+          label: String(option),
+          disabled: false,
+          origin,
+        };
 
     // Duplicate value is no longer added
     if (optionInfoMap.get(optionInfo.value)) {
@@ -142,47 +112,7 @@ export const getOptionNodes = ({
     }
   };
 
-  const travelChildren = (children: VNode[]) => {
-    for (const child of children) {
-      if (isNamedComponent(child, 'Optgroup')) {
-        // OptGroup处理
-        nodes.push({
-          type: 'optGroup',
-          key: `group-${child.props?.label}`,
-          label: child.props?.label,
-        });
-        // 添加Group下面的Options
-        if (isSlotsChildren(child, child.children)) {
-          const _children = child.children.default?.() ?? [];
-          travelChildren(_children);
-        } else if (isArrayChildren(child, child.children)) {
-          travelChildren(child.children);
-        }
-      } else if (isNamedComponent(child, 'Option')) {
-        // Option处理
-        const optionInfo = getAndSaveOptionInfo(child, 'children');
-        if (optionInfo && isValidOption(optionInfo)) {
-          nodes.push({
-            type: 'option',
-            key: optionInfo.key,
-            value: optionInfo.value,
-            label: optionInfo.label,
-            render: optionInfo.render,
-            disabled: optionInfo.disabled,
-          });
-          if (!optionInfo.disabled) {
-            enabledOptionSet.add(optionInfo.value);
-          }
-        }
-      } else if (isArrayChildren(child, child.children)) {
-        travelChildren(child.children);
-      }
-    }
-  };
-
-  if (children) {
-    travelChildren(children);
-  } else if (options) {
+  if (options) {
     extendChildrenFromOptions(options, 'options');
   }
   if (extraOptions) {
