+++ packages/web-vue/components/_hooks/use-options.tsx
@@ -1,14 +1,14 @@
-import { ComponentPublicInstance, computed, Ref, ref, Slot, watch } from 'vue';
+import { ComponentPublicInstance, Ref, ref, watch } from 'vue';
 import { getRelativeRect } from '../_utils/dom';
 import {
   FilterOption,
   Option,
   OptionInfo,
+  OptionNode,
 } from '../_components/dropdown/interface';
 import { getOptionNodes } from '../_components/dropdown/utils';
 
 export const useOptions = ({
-  defaultSlot,
   options,
   extraOptions,
   inputValue,
@@ -17,7 +17,6 @@ export const useOptions = ({
   optionRefs,
   virtualListRef,
 }: {
-  defaultSlot?: Ref<Slot | undefined>;
   options?: Ref<Option[]>;
   extraOptions?: Ref<Array<string | number>>;
   inputValue?: Ref<string>;
@@ -73,12 +72,10 @@ export const useOptions = ({
     }
   };
 
-  const optionNodes = computed(() => {
-    optionInfoMap.clear();
-    enabledOptionSet.clear();
+  const nodes = ref<OptionNode[]>([]);
 
-    return getOptionNodes({
-      children: defaultSlot?.value?.(),
+  const setOptionNodes = () => {
+    nodes.value = getOptionNodes({
       options: options?.value,
       extraOptions: extraOptions?.value,
       inputValue: inputValue?.value,
@@ -86,24 +83,18 @@ export const useOptions = ({
       optionInfoMap,
       enabledOptionSet,
     });
-  });
+  };
 
-  // When the option changes, recalculate the activeOption
   watch(
-    optionNodes,
+    [options, extraOptions, inputValue, filterOption],
     () => {
-      if (
-        enabledOptionSet.size > 0 &&
-        (!activeOption.value || !enabledOptionSet.has(activeOption.value.value))
-      ) {
-        activeOption.value = optionInfoMap.get(Array.from(enabledOptionSet)[0]);
-      }
+      setOptionNodes();
     },
     { immediate: true }
   );
 
   return {
-    optionNodes,
+    nodes,
     optionInfoMap,
     enabledOptionSet,
     activeOption,
